+++ src/extension.ts
@@ -29,6 +29,7 @@ import { DocumentHighlightProvider } from "./providers/DocumentHighlightProvider
 import { WorkspaceSymbolProvider } from "./providers/WorkspaceSymbolProvider";
 import { FoldingRangeProvider } from "./providers/FoldingRangeProvider";
 import { ColorProvider } from "./providers/ColorProvider";
+import { QB64EvaluatableExpressionProvider } from "./providers/EvaluatableExpressionProvider";
 import { IndexDiagnostics } from "./providers/IndexDiagnostics";
 import { CallHierarchyProvider } from "./providers/CallHierarchyProvider";
 import {
@@ -230,6 +231,13 @@ export async function activate(context: vscode.ExtensionContext) {
 
   context.subscriptions.push(
     vscode.languages.registerColorProvider(documentSelector, new ColorProvider())
+  );
+
+  context.subscriptions.push(
+    vscode.languages.registerEvaluatableExpressionProvider(
+      documentSelector,
+      new QB64EvaluatableExpressionProvider()
+    )
   );
 
   const semanticTokensProvider = new SemanticTokensProvider(workspaceIndex);